Create the backtesting algorithm
This new backtesting algorithm is based on the Financialmodelingprep API (instead of stocksrow.com). I am not finished with the backtesting algorithm but will soon. I still need to put the 3 files into 1 umbrella function and write the code to string interpolation, as well as make some changes to final_two.py file where I will need to show whether there was a 100% gain within the 3 year period. And if not, whether there was a capital gain or capital loss between the buy_share_price and the sell_share_price